+++ ChangeLog.md
@@ -1,5 +1,5 @@
 # ChangeLog
 
 ## v0.0.1 (2014-03-21)
-* Added namespace
 * Added Changelog
+* Made vsRegions dynamic